More hacking on things.
- Sanely handle build phase for :all targets (i.e. nothing)
- Sane error if asked to build something that doesn't exist
- Better handle git file-not-found errors
- Skip the git fetch all; just fetch the ref we actually want!
- Preliminary implementation of metahashing of build input factors for
  caching and build invalidation
- Disallow virtual targets with no deps

This is able to build mini-boot now, as long as you list out all its
srcs. Any genrule should work, with a caveat around collecting deps not
being done correctly yet.  They'll tend to work for now just because
the deps will still exist in the build tree when the thing that needs
them runs.