Dedupe cached objects by their sha1 checksum.
If a build run puts a new version of a generated file that
theoretically should not have changed (had the build been truly
hermetic), store the new version and emit a warning. For the time being
this will be a common occurrence in cases such as gendeb, wherein the
output file has a timestamp embedded in it.

When replacing an existing object with a new version, the old one still
remains in the hashed-by-file's-checksum directory structure. They
could accumulate over time, so perhaps I should add a method to
periodically prune unreferenced objects from the cache.

This should interact reasonably well with a "store build artifacts
forever" feature later on.